+++ source/Attenuator/Header.h
@@ -55,7 +55,7 @@ const uint8_t i_fast_led_delay = 3;
  * Manage the color and blink pattern for the top LED.
  */
 millisDelay ms_top_blink; // Allows the top LED to blink for a menu state.
-const unsigned int i_top_blink_delay = 800; // Duration for blink pattern.
+const uint16_t i_top_blink_delay = 800; // Duration for blink pattern.
 uint8_t i_top_led_color; // Remember the last color for the top LED.
 uint8_t i_top_led_brightness = 128; // Max brightness for this LED.
 bool b_top_led_off = false; // Denotes when top LED is mid-blink.
@@ -88,15 +88,15 @@ bool b_buzzer_on = false;
 bool b_vibrate_on = false;
 const uint8_t i_min_power = 0;
 const uint8_t i_max_power = 255;
-const unsigned int i_buzzer_max_time = 300; // Longest duration for a standalone "beep".
-const unsigned int i_vibrate_min_time = 500; // Minimum runtime for vibration motor.
-const unsigned int i_vibrate_max_time = 1500; // Maximum runtime for vibration motor.
+const uint16_t i_buzzer_max_time = 300; // Longest duration for a standalone "beep".
+const uint16_t i_vibrate_min_time = 500; // Minimum runtime for vibration motor.
+const uint16_t i_vibrate_max_time = 1500; // Maximum runtime for vibration motor.
 
 /*
  * For the alarm and venting/overheat, set the blink/buzz/vibrate interval.
  */
 millisDelay ms_blink_leds;
-const unsigned int i_blink_leds = 600;
+const uint16_t i_blink_leds = 600;
 bool b_blink_blank = false; // Denotes when upper/lower LEDs are mid-blink.
 
 /*
@@ -208,8 +208,8 @@ millisDelay ms_rotary_debounce; // Put some timing on the rotary so we do not ov
 millisDelay ms_center_double_tap; // Timer for determinine when a double-tap was detected.
 millisDelay ms_center_long_press; // Timer for determining when a long press was detected.
 bool b_center_pressed = false;
-const unsigned int i_center_double_tap_delay = 300; // When to consider the center dial has a "double tap".
-const unsigned int i_center_long_press_delay = 600; // When to consider the center dial has a "long" press.
+const uint16_t i_center_double_tap_delay = 300; // When to consider the center dial has a "double tap".
+const uint16_t i_center_long_press_delay = 600; // When to consider the center dial has a "long" press.
 uint8_t i_press_count = 0;
 uint8_t i_rotary_count = 0;
 int i_encoder_pos = 0;
@@ -227,11 +227,11 @@ enum MENU_LEVELS MENU_LEVEL;
 /*
 * Music Track Info and Playback States
 */
-const unsigned int i_music_track_offset = 500; // Music tracks always start at index 500.
-unsigned int i_music_track_count = 0; // Count of tracks as returned by the pack.
-unsigned int i_music_track_current = 0;
-unsigned int i_music_track_min = 0; // Min value for music track index (0 = unset).
-unsigned int i_music_track_max = 0; // Max value for music track index (0 = unset).
+const uint16_t i_music_track_offset = 500; // Music tracks always start at index 500.
+uint16_t i_music_track_count = 0; // Count of tracks as returned by the pack.
+uint16_t i_music_track_current = 0;
+uint16_t i_music_track_min = 0; // Min value for music track index (0 = unset).
+uint16_t i_music_track_max = 0; // Max value for music track index (0 = unset).
 uint8_t i_volume_master_percentage = 100; // Master overall volume
 uint8_t i_volume_effects_percentage = 100; // Sound effects
 uint8_t i_volume_music_percentage = 100; // Music volume
